Refactored util.unpack package to util.preload
