+++ src/main/java/com/github/otbproject/otbproject/App.java
@@ -17,8 +17,8 @@ import com.github.otbproject.otbproject.util.LibsLoader;
 import com.github.otbproject.otbproject.util.UnPacker;
 import com.github.otbproject.otbproject.util.VersionClass;
 import com.github.otbproject.otbproject.util.compat.VersionCompatHelper;
-import com.github.otbproject.otbproject.util.unpack.LoadStrategy;
-import com.github.otbproject.otbproject.util.unpack.PreloadLoader;
+import com.github.otbproject.otbproject.util.preload.LoadStrategy;
+import com.github.otbproject.otbproject.util.preload.PreloadLoader;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.ParseException;
 import org.apache.logging.log4j.LogManager;
