+++ src/main/java/com/github/otbproject/otbproject/cli/commands/CmdParser.java
@@ -10,8 +10,8 @@ import com.github.otbproject.otbproject.messages.internal.InternalMessageSender;
 import com.github.otbproject.otbproject.messages.receive.PackagedMessage;
 import com.github.otbproject.otbproject.messages.send.MessagePriority;
 import com.github.otbproject.otbproject.users.UserLevel;
-import com.github.otbproject.otbproject.util.unpack.LoadStrategy;
-import com.github.otbproject.otbproject.util.unpack.PreloadLoader;
+import com.github.otbproject.otbproject.util.preload.LoadStrategy;
+import com.github.otbproject.otbproject.util.preload.PreloadLoader;
 import com.google.common.collect.ImmutableSet;
 
 import java.awt.*;
